% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_execucao.R
\name{baixar_execucao}
\alias{baixar_execucao}
\title{Baixar execução penal do TJRJ}
\usage{
baixar_execucao(usuario = NULL, senha = NULL, processo = NULL, diretorio = ".")
}
\arguments{
\item{usuario}{usuário}

\item{senha}{senha}

\item{processo}{Número do processo da execução penal}

\item{diretorio}{Diretório onde colocar os htmls}
}
\value{
html da execução, html da parte, htmls dos processos
criminais
}
\description{
Baixar execução penal do TJRJ
}
