% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_execucao1.R
\name{baixar_execucao1}
\alias{baixar_execucao1}
\title{Baixar execução penal do TJRJ (apenas execucao)}
\usage{
baixar_execucao1(
  usuario = NULL,
  senha = NULL,
  processo = NULL,
  diretorio = "."
)
}
\arguments{
\item{usuario}{usuário}

\item{senha}{senha}

\item{processo}{Número do processo da execução penal}

\item{diretorio}{Diretório onde colocar os htmls}
}
\value{
html da execução
}
\description{
Baixar execução penal do TJRJ (apenas execucao)
}
