% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_execucao_completa.R
\name{baixar_execucao_completa}
\alias{baixar_execucao_completa}
\title{Baixa html, pdfs e tabela com indice de docs}
\usage{
baixar_execucao_completa(
  usuario = NULL,
  senha = NULL,
  processo = NULL,
  diretorio = "."
)
}
\arguments{
\item{usuario}{Usuário}

\item{senha}{Senha}

\item{diretorio}{Diretório. Default para atual.}
}
\value{
htmls, pdfs e rds
}
\description{
Baixa html, pdfs e tabela com indice de docs
}
