% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_numeros.R
\name{baixar_numeros}
\alias{baixar_numeros}
\title{Baixa números dos processos}
\usage{
baixar_numeros(
  login = NULL,
  passwd = NULL,
  start = NULL,
  end = NULL,
  diretorio = "."
)
}
\arguments{
\item{login}{Login do usuário previamente cadastrado}

\item{passwd}{Senha do usuário previamente cadastrado}

\item{start}{Data inicial no formato dd/mm/aaaa}

\item{end}{Data final no formato dd/mm/aaaa}

\item{diretorio}{Diretório onde armazenar os arquivos}
}
\value{
htmls
}
\description{
Baixa números dos processos
}
\details{
Esta função é interna. Ela pressupõe que esteja com o
conteiner do Selenium rodando e o RSelenium carregado.
}
