% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_numeros_execucao.R
\name{baixar_numeros_execucao}
\alias{baixar_numeros_execucao}
\title{Baixar números das execuções}
\usage{
baixar_numeros_execucao(
  usuario = NULL,
  senha = NULL,
  data_inicial = NULL,
  data_final = NULL,
  diretorio = "."
)
}
\arguments{
\item{usuario}{Usuário}

\item{senha}{Senha}

\item{data_inicial}{Data inicial no formato dd/mm/yyyy}

\item{data_final}{Data final no formato dd/mm/yyyy}

\item{diretorio}{Diretório. Default para atual.}
}
\value{
htmls
}
\description{
Baixar números das execuções
}
