% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pdf.R
\name{baixar_pdf}
\alias{baixar_pdf}
\title{Baixa pdfs}
\usage{
baixar_pdf(usuario = NULL, senha = NULL, id_proc = NULL, arquivo = NULL)
}
\arguments{
\item{usuario}{Usuário}

\item{senha}{Senha}

\item{id_proc}{id_execucao}

\item{arquivo}{arquivo}
}
\value{
pdf
}
\description{
Baixa pdfs
}
