% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pdf2.R
\name{baixar_pdf2}
\alias{baixar_pdf2}
\title{Baixa pdfs}
\usage{
baixar_pdf2(usuario = NULL, senha = NULL, processo = NULL, diretorio = ".")
}
\arguments{
\item{usuario}{Usuário}

\item{senha}{Senha}

\item{processo}{Número do processo}

\item{diretorio}{Onde colocar os pdfs}
}
\value{
pdf
}
\description{
Baixa pdfs
}
