% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endereco_sentenciado.R
\name{endereco_sentenciado}
\alias{endereco_sentenciado}
\title{Endereco do sentenciado}
\usage{
endereco_sentenciado(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar onde se encontram os arquivos, se param arquivos for NULL}
}
\value{
tibble
}
\description{
Endereco do sentenciado
}
