% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execucaopenal.R
\docType{package}
\name{execucaopenal-package}
\alias{execucaopenal}
\alias{execucaopenal-package}
\title{\code{execucaopenal} package}
\description{
Baixa  e organiza execução penal do Rio de Janeiro
}
\author{
\strong{Maintainer}: Jose de Jesus Filho \email{jose@consudata.com.br} (\href{https://orcid.org/0000-0001-7415-7383}{ORCID})

}
