% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_carta_sentenca.R
\name{extrair_carta_sentenca}
\alias{extrair_carta_sentenca}
\title{Extrair dados da carta de sentença}
\usage{
extrair_carta_sentenca(strings = NULL)
}
\arguments{
\item{strings}{carta de sentença lido com pdftools}
}
\value{
tibble
}
\description{
Extrair dados da carta de sentença
}
