% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_penas.R
\name{extrair_penas}
\alias{extrair_penas}
\title{Extrair penas dos atestados}
\usage{
extrair_penas(indice = NULL)
}
\arguments{
\item{indice}{data.frame extraído com ler_indice}
}
\value{
tibble
}
\description{
Extrair penas dos atestados
}
