% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_qualificacao.R
\name{extrair_qualificacao}
\alias{extrair_qualificacao}
\title{Extrair qualificação do atestado de pena}
\usage{
extrair_qualificacao(indice)
}
\arguments{
\item{indice}{dataframe extraído com ler_índice}
}
\value{
tibble
}
\description{
Extrair qualificação do atestado de pena
}
