% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generos.R
\docType{data}
\name{generos}
\alias{generos}
\title{generos}
\format{
A data frame with 8 rows and 2 variables:
\describe{
\item{\code{codigo}}{character código do gênero do preso}
\item{\code{genero}}{character descrição do gênero do preso}
}
}
\source{
\url{https://seeu.pje.jus.br/seeu/}
}
\usage{
generos
}
\description{
Código do gênero do preso para usar na função baixar_numeros_execucao
}
\keyword{datasets}
