% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_conhecimento.R
\name{ler_conhecimento}
\alias{ler_conhecimento}
\title{Ler conhecimento}
\usage{
ler_conhecimento(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{arquivos}

\item{diretorio}{diretorio}
}
\value{
tibble
}
\description{
Ler conhecimento
}
