% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_doc_indice.R
\name{ler_doc_indice}
\alias{ler_doc_indice}
\title{Constrói índice das páginas dos pdfs baixados}
\usage{
ler_doc_indice(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório de arquivos, se arquivos for NULL}
}
\value{
tibble
}
\description{
Constrói índice das páginas dos pdfs baixados
}
