% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_execucao.R
\name{ler_execucao}
\alias{ler_execucao}
\title{Ler dados da execução penal}
\usage{
ler_execucao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar diretório, se arquivos for NULL}
}
\value{
tibble
}
\description{
Ler dados da execução penal
}
