% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_execucao2.R
\name{ler_execucao2}
\alias{ler_execucao2}
\title{Ler execucao2}
\usage{
ler_execucao2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Arquivos}

\item{diretorio}{Diretório}
}
\value{
tibble
}
\description{
Ler execucao2
}
