% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao.R
\name{ler_movimentacao}
\alias{ler_movimentacao}
\title{Ler movimentação processual da execução penal}
\usage{
ler_movimentacao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório}
}
\value{
tibble
}
\description{
Ler movimentação processual da execução penal
}
