% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_numeros.R
\name{ler_numeros}
\alias{ler_numeros}
\title{Ler números dos processos baixas a partir do script /Inst/baixar_numeros}
\usage{
ler_numeros(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Inforrmar diretório se arquivos for NULL}
}
\value{
tibble
}
\description{
Ler números dos processos baixas a partir do script /Inst/baixar_numeros
}
