% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_partes.R
\name{ler_partes}
\alias{ler_partes}
\title{Ler partes}
\usage{
ler_partes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Arquivos}

\item{diretorio}{Diretório}
}
\value{
tibble
}
\description{
Ler partes
}
