% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_perfil.R
\name{ler_perfil}
\alias{ler_perfil}
\title{Extrai perfil do sentenciado}
\usage{
ler_perfil(arquivos = NULL, diretorio = ".", organizar = TRUE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se arquivos for NULL}

\item{organizar}{Exclui colunas pouco relevantes e transforma
dados, normalizando nomes e padronizando datas}
}
\value{
tibble
}
\description{
Extrai perfil do sentenciado
}
\details{
O argumento organizar pode provocar erro se forem lidos um ou poucos
processos, pois nem todas as colunas aparecem. Assim, para ler um ou dois
ajuste para FALSE. No entanto, os dados não estarão bem padronizados.
}
