% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_processo_criminal.R
\name{ler_processo_criminal}
\alias{ler_processo_criminal}
\title{Processos criminais dos sentenciados}
\usage{
ler_processo_criminal(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Arquivos a serem lidos}

\item{diretorio}{Diretório, se arquivos for NULL}
}
\value{
tibble
}
\description{
Processos criminais dos sentenciados
}
