% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/situacao_preso.R
\docType{data}
\name{situacao_preso}
\alias{situacao_preso}
\title{situacao_preso}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{\code{codigo}}{character código da situação do presos}
\item{\code{situacao}}{character descrição da situação do preso}
}
}
\source{
\url{https://seeu.pje.jus.br/seeu/}
}
\usage{
situacao_preso
}
\description{
Código da situação do preso para usar na função baixar_numeros_execucao
}
\keyword{datasets}
