% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribunais.R
\docType{data}
\name{tribunais}
\alias{tribunais}
\title{tribunais}
\format{
A data frame with 76 rows and 2 variables:
\describe{
\item{\code{codigo}}{character código da situação do presos}
\item{\code{tribunal}}{character descrição do tribunal}
}
}
\source{
\url{https://seeu.pje.jus.br/seeu/}
}
\usage{
tribunais
}
\description{
Código do tribunal para usar na função baixar_numeros_execucao
}
\keyword{datasets}
