% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pdf.R
\name{baixar_pdf}
\alias{baixar_pdf}
\title{Baixa pdfs}
\usage{
baixar_pdf(usuario, senha, id_proc, arquivo)
}
\arguments{
\item{usuario}{Usuário}

\item{senha}{Senha}

\item{id_proc}{id_execucao}

\item{arquivo}{arquivo}
}
\value{
pdf
}
\description{
Baixa pdfs
}
