% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pdf.R
\name{baixar_pdf}
\alias{baixar_pdf}
\title{Baixa pdfs}
\usage{
baixar_pdf(
  usuario = NULL,
  senha = NULL,
  id_proc = NULL,
  processo = NULL,
  arquivo = NULL
)
}
\arguments{
\item{usuario}{Usuário}

\item{senha}{Senha}

\item{id_proc}{Id do processo}

\item{processo}{Número do processo}

\item{arquivo}{arquivo}
}
\value{
pdf
}
\description{
Baixa pdfs
}
